requirements.txt
dateparser
matplotlib
numpy
pyModbusTCP
pytz
Requests
scipy
zipfile
Pillow